+++ srcs/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: nforay <nforay@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/12/01 17:23:59 by nforay            #+#    #+#             */
-/*   Updated: 2020/02/01 16:20:17 by nforay           ###   ########.fr       */
+/*   Updated: 2020/02/02 14:43:35 by nforay           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 
 	if (!s)
 		return (NULL);
-	if (ft_strlen(s) < (int)start)
+	if (ft_strlen(s) < start)
 		return (ft_strdup(""));
 	if (!(ptr = malloc(len + 1)))
 		return (NULL);
